feat: route audio-only publications to the audio reader host on iOS
readerViewKind now has an audio case, checked after PDF and image so
their precedence is unchanged, and the factory builds AudioReaderView
for it. An audio-only publication no longer mounts an EPUB navigator it
has nothing to render into.

That mattered beyond the wasted work. The navigator's preloaded spreads
fetch the reading-order hrefs, so for a streamed audiobook the WKWebViews
were pulling the same URL AVFoundation was streaming, and two iOS
integration tests failed once the publication-swap fix made it happen
twice per test. Both pass again.

The predicate is Readium's own conforms(to: .audiobook), which is
readingOrder.allAreAudio behind a non-empty guard. metadata.conformsTo
is deliberately not consulted here: a media-overlay EPUB can declare the
audiobook profile over an HTML reading order, and that publication has
pages to render.

ReaderViewKindRoutingTests moved to its own file. The six new cases
pushed ReadiumExtensionsTests past the file-size limit, and it held four
unrelated test classes.
