+++ flureadium/example/ios/RunnerTests/ReadiumExtensionsTests.swift
@@ -85,96 +85,6 @@ final class LocatorParsingTests: XCTestCase {
     }
 }
 
-final class ReaderViewKindRoutingTests: XCTestCase {
-
-    func testReturnsPdfForPdfPublication() {
-        let publication = publication(
-            profile: .pdf,
-            readingOrder: [link(href: "/chapter.xhtml", type: .xhtml)]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .pdf)
-    }
-
-    func testReturnsImageForDivinaPublication() {
-        let publication = publication(
-            profile: .divina,
-            readingOrder: [link(href: "/chapter.xhtml", type: .xhtml)]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .image)
-    }
-
-    func testReturnsImageForBitmapReadingOrder() {
-        let publication = publication(
-            profile: nil,
-            readingOrder: [
-                link(href: "/page-1.jpg", type: .jpeg),
-                link(href: "/page-2.png", type: .png),
-            ]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .image)
-    }
-
-    func testReturnsPdfWhenPublicationMatchesPdfAndImageSignals() {
-        let publication = publication(
-            profile: .pdf,
-            readingOrder: [
-                link(href: "/page-1.jpg", type: .jpeg),
-                link(href: "/page-2.png", type: .png),
-            ]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .pdf)
-    }
-
-    func testReturnsImageForSingleBitmapResource() {
-        let publication = publication(
-            profile: nil,
-            readingOrder: [link(href: "/page-1.jpg", type: .jpeg)]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .image)
-    }
-
-    func testReturnsEpubForMixedBitmapAndHtmlReadingOrder() {
-        let publication = publication(
-            profile: nil,
-            readingOrder: [
-                link(href: "/page-1.jpg", type: .jpeg),
-                link(href: "/chapter.xhtml", type: .xhtml),
-            ]
-        )
-
-        XCTAssertEqual(readerViewKind(for: publication), .epub)
-    }
-
-    func testReturnsEpubForEmptyReadingOrder() {
-        let publication = publication(profile: nil, readingOrder: [])
-
-        XCTAssertEqual(readerViewKind(for: publication), .epub)
-    }
-
-    private func publication(profile: Publication.Profile?, readingOrder: [Link]) -> Publication {
-        var conformsTo: [Publication.Profile] = []
-        if let profile {
-            conformsTo = [profile]
-        }
-
-        return Publication(
-            manifest: Manifest(
-                metadata: Metadata(conformsTo: conformsTo, title: "Reader Kind"),
-                readingOrder: readingOrder
-            )
-        )
-    }
-
-    private func link(href: String, type: MediaType) -> Link {
-        Link(href: href, mediaType: type)
-    }
-}
-
 // MARK: - State Mapping Tests
 
 final class StateMappingTests: XCTestCase {
